Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float num1;
-    scanf("%f",&num1);
-    printf("You entered: %.4f",num1);
+    double num1;
+    scanf("%lf",&num1);
+    printf("You entered: %.4lf",num1);
     return 0;
 }
